+++ src/lib/components/genomics/DnaTranslator.svelte
