+++ src/lib/components/genomics/FastaStats.svelte
